+++ packages/picasso/CHANGELOG.md
@@ -1,5 +1,54 @@
 # Change Log
 
+## 21.2.0
+
+### Minor Changes
+
+- [#2588](https://github.com/toptal/picasso/pull/2588) [`ace2ae7c`](https://github.com/toptal/picasso/commit/ace2ae7c1e2771fd87985e0e62eaf8dc71d3c424) Thanks [@ozgurkececioglu](https://github.com/ozgurkececioglu)! - ---
+
+  ### RichTextEditor
+
+  - added new prop `status` to indicate the status of input field
+  - added deprecation warning for `error` prop usage
+
+### Patch Changes
+
+- [#2589](https://github.com/toptal/picasso/pull/2589) [`e6ee5341`](https://github.com/toptal/picasso/commit/e6ee5341d2fd75659733f2a91b21a80912840e60) Thanks [@sunRock98](https://github.com/sunRock98)! - ---
+
+  ### DatePicker
+
+  - fixed a bug, that disabled DatePicker when clicked was showing a Calendar
+
+## 21.1.0
+
+### Minor Changes
+
+- [#2547](https://github.com/toptal/picasso/pull/2547) [`070afaeb`](https://github.com/toptal/picasso/commit/070afaebd68da0c42310e0757aa04b7f0d716257) Thanks [@ascrazy](https://github.com/ascrazy)! - ---
+
+  ### DatePicker
+
+  - add `popperProps` prop to let consumers control data-\* attrs of the Popper
+
+### Patch Changes
+
+- [#2597](https://github.com/toptal/picasso/pull/2597) [`7b6d3382`](https://github.com/toptal/picasso/commit/7b6d3382c67f11780dfcb796728ee25ce3273803) Thanks [@TomasSlama](https://github.com/TomasSlama)! - Move `storybook-readme` to devDependencies to reduce bundle size
+
+## 21.0.0
+
+### Major Changes
+
+- [#2569](https://github.com/toptal/picasso/pull/2569) [`e8833df2`](https://github.com/toptal/picasso/commit/e8833df2329c6285840ae02c0d029fe4eebb8247) Thanks [@augustobmoura](https://github.com/augustobmoura)! - ---
+
+  ### Project
+
+  - Update typescript to version 4.6
+
+### Patch Changes
+
+- Updated dependencies [[`e8833df2`](https://github.com/toptal/picasso/commit/e8833df2329c6285840ae02c0d029fe4eebb8247)]:
+  - @toptal/picasso-provider@1.0.0
+  - @toptal/picasso-shared@8.0.0
+
 ## 20.1.0
 
 ### Minor Changes